(PC-2555): BSR -> re-organized activation, forgot-password folder
