+++ src/components/pages/activation/Error/__specs__/Error.spec.jsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { shallow } from 'enzyme'
 
 import Error from '../Error'
-import MailToLink from '../../../layout/MailToLink/MailToLink'
-import { SUPPORT_EMAIL, SUPPORT_EMAIL_SUBJECT } from '../../../../utils/config'
+import MailToLink from '../../../../layout/MailToLink/MailToLink'
+import { SUPPORT_EMAIL, SUPPORT_EMAIL_SUBJECT } from '../../../../../utils/config'
 
 describe('src | components | pages | activation | Error', () => {
   it('should match the snapshot', () => {
