+++ src/components/pages/forgot-password/RequestEmailForm/__specs__/RequestEmailForm.spec.jsx
@@ -1,14 +1,14 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import { ResetThePasswordForm } from '../ResetThePasswordForm'
+import { RequestEmailForm } from '../RequestEmailForm'
 
-describe('src | components | pages | forgot-password | ResetThePasswordForm', () => {
+describe('src | components | pages | forgot-password | RequestEmailForm', () => {
   it('should match the snapshot', () => {
     // given
     const props = { canSubmit: true, isLoading: false }
 
     // when
-    const wrapper = shallow(<ResetThePasswordForm {...props} />)
+    const wrapper = shallow(<RequestEmailForm {...props} />)
 
     // then
     expect(wrapper).toMatchSnapshot()
