+++ src/components/pages/forgot-password/SuccessView/__specs__/SuccessView.spec.jsx
