+++ src/components/pages/forgot-password/hocs/withResetForm.jsx
@@ -4,7 +4,8 @@ import { Form as FinalForm } from 'react-final-form'
 import { connect } from 'react-redux'
 import { requestData } from 'redux-thunk-data'
 
-const noop = () => {}
+const noop = () => {
+}
 
 const withPasswordForm = (WrappedComponent, validator, routePath, routeMethod) => {
   const name = WrappedComponent.displayName || 'Component'
@@ -58,14 +59,14 @@ const withPasswordForm = (WrappedComponent, validator, routePath, routeMethod) =
     }
 
     renderFinalForm = ({
-      // https://github.com/final-form/final-form#formstate
-      dirtySinceLastSubmit,
-      error: preSubmitError,
-      handleSubmit,
-      hasSubmitErrors,
-      hasValidationErrors,
-      pristine,
-    }) => {
+                         // https://github.com/final-form/final-form#formstate
+                         dirtySinceLastSubmit,
+                         error: preSubmitError,
+                         handleSubmit,
+                         hasSubmitErrors,
+                         hasValidationErrors,
+                         pristine,
+                       }) => {
       const { isloading } = this.state
       const canSubmit =
         (!pristine && !hasSubmitErrors && !hasValidationErrors && !isloading) ||
